remove function from public api
