+++ libs/core/src/graphql/build-fragment-name-spread.ts
@@ -1,6 +1,19 @@
-import { DocumentNode } from 'graphql'
+import { DocumentNode, FragmentDefinitionNode } from 'graphql'
 
-import { daffGetFragmentNames } from './get-fragment-names'
+const getFragmentNames = (fragment: DocumentNode) =>
+  fragment.definitions.filter(def =>
+    def.kind === 'FragmentDefinition'
+  ).map(def =>
+    (def as FragmentDefinitionNode).name.value
+  )
+
+/**
+ * Builds a list of fragment names present inside the specified GraphQL document nodes.
+ * Returns an empty array if no fragments have been defined or if null is passed.
+ * @param fragments The created fragments.
+ */
+const daffGetFragmentNames = (...fragments: DocumentNode[]): string[] =>
+  fragments.reduce((acc, fragment) => acc.concat(getFragmentNames(fragment)), [])
 
 /**
  * Builds a string of fragment names that can be interpolated into a GraphQL query.
