+++ libs/core/src/graphql/public_api.ts
@@ -1,3 +1,2 @@
-export { daffGetFragmentNames } from './get-fragment-names';
 export { daffBuildFragmentNameSpread } from './build-fragment-name-spread';
 export { daffBuildFragmentDefinition } from './build-fragment-definition';
